+++ src/entities/postsManager/hooks/index.ts
@@ -1,0 +1,1 @@
+export { usePostActions } from "./usePostActions"
